# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Filmophilia - A movie discovery and social platform. JSON endpoints return `{success,data}` on success and `{success,error}` on failure.",
        "title": "Filmophilia API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@filmophilia.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/me/watchlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the authenticated user's watchlist with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Get my watchlist",
                "parameters": [
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Filter: all, watched, unwatched",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Results per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watchlist items",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.WatchlistItemResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/watchlist/count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get count of movies in the authenticated user's watchlist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Get watchlist count",
                "responses": {
                    "200": {
                        "description": "Watchlist count",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.WatchlistCountResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/movies/{movieId}/watchlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific movie's watchlist entry for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Get watchlist entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Watchlist entry",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.WatchlistEntryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid movie ID",
                        "schema": {
                            "$ref": "#/definitions/dto.BadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Movie not in watchlist",
                        "schema": {
                            "$ref": "#/definitions/dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a movie to the authenticated user's watchlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Add movie to watchlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional notes and rank",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.AddToWatchlistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Movie added to watchlist",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.WatchlistEntryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.BadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "schema": {
                            "$ref": "#/definitions/dto.NotFoundErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Movie already in watchlist",
                        "schema": {
                            "$ref": "#/definitions/dto.ConflictErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a movie from the authenticated user's watchlist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Remove from watchlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.MessageData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid movie ID",
                        "schema": {
                            "$ref": "#/definitions/dto.BadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Movie not in watchlist",
                        "schema": {
                            "$ref": "#/definitions/dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update notes or rank for a movie in the watchlist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Update watchlist entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateWatchlistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated entry",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.WatchlistEntryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.BadRequestErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Movie not in watchlist",
                        "schema": {
                            "$ref": "#/definitions/dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/movies/{movieId}/watchlist/check": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check if a movie is in the authenticated user's watchlist",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watchlist"
                ],
                "summary": "Check if movie is in watchlist",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Movie ID",
                        "name": "movieId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In watchlist status",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.WatchlistStatusData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid movie ID",
                        "schema": {
                            "$ref": "#/definitions/dto.BadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddToWatchlistRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string"
                },
                "rank": {
                    "type": "number"
                }
            }
        },
...
        "dto.BadRequestErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.BadRequestError"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.ConflictErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ConflictErrorInfo"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.InternalServerErrorInfo"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.MessageData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "operation completed successfully"
                }
            }
        },
...
        "dto.NotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.NotFoundErrorInfo"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.UpdateWatchlistRequest": {
            "type": "object",
            "properties": {
                "notes": {
                    "type": "string"
                },
                "rank": {
                    "type": "number"
                }
            }
        },
...
        "dto.WatchlistCountResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "unwatched": {
                    "type": "integer"
                },
                "watched": {
                    "type": "integer"
                }
            }
        },
        "dto.WatchlistEntryResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "movie_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "rank_position": {
                    "type": "number"
                },
                "watched_at": {
                    "type": "string"
                }
            }
        },
        "dto.WatchlistItemResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "movie_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "poster_url": {
                    "type": "string"
                },
                "rank_position": {
                    "type": "number"
                },
                "release_date": {
                    "type": "string"
                },
                "runtime": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_avg_rating": {
                    "type": "number"
                },
                "watched_at": {
                    "type": "string"
                }
            }
        },
        "dto.WatchlistStatusData": {
            "type": "object",
            "properties": {
                "in_watchlist": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}